add query cancel option for postgres only
by using qt classes for db, seems that this will be vendor specific.
first implementation (test?) is done for postgres and probably needs to
be refactored.

there is an error written in console which i'm not sure if has to be
addressed or not (might be that there is a mem leak)
error is: "Unable to free statement: connection pointer is NULL"